Auction-App-Data-Processing_AWS-Pipeline/tests/integration/test_pipeline.py
"""
Integration Tests for Auction Pipeline

Tests the end-to-end functionality of pipeline components.
Requires Docker services to be running.
"""

import json
import os
import time
from typing import List

import pytest


# Skip all tests if Docker services aren't available
pytestmark = pytest.mark.integration


class TestKafkaIntegration:
    """Integration tests for Kafka producer/consumer."""
    
    @pytest.mark.kafka
    def test_produce_and_consume_bid(self, kafka_bootstrap_servers: str):
        """Test producing and consuming a bid event."""
        pytest.importorskip("kafka")
        
        from kafka import KafkaProducer, KafkaConsumer
        
        topic = "test.auction.bids"
        
        # Create producer
        producer = KafkaProducer(
            bootstrap_servers=kafka_bootstrap_servers,
            value_serializer=lambda v: json.dumps(v).encode('utf-8'),
        )
        
        # Send test message
        test_bid = {
            "bid_id": "test-123",
            "auction_id": "auction-456",
            "bid_amount": 100.00,
        }
        
        future = producer.send(topic, test_bid)
        result = future.get(timeout=10)
        
        assert result.topic == topic
        assert result.partition >= 0
        
        # Create consumer
        consumer = KafkaConsumer(
            topic,
            bootstrap_servers=kafka_bootstrap_servers,
            auto_offset_reset='earliest',
            value_deserializer=lambda m: json.loads(m.decode('utf-8')),
            consumer_timeout_ms=5000,
        )
        
        # Consume message
        messages = []
        for msg in consumer:
            messages.append(msg.value)
            if msg.value.get("bid_id") == "test-123":
                break
        
        assert len(messages) > 0
        assert any(m.get("bid_id") == "test-123" for m in messages)
        
        producer.close()
        consumer.close()


class TestPostgresIntegration:
    """Integration tests for PostgreSQL operations."""
    
    @pytest.mark.postgres
    def test_database_connection(self, postgres_connection_params: dict):
        """Test PostgreSQL connection."""
        pytest.importorskip("psycopg2")
        
        import psycopg2
        
        conn = psycopg2.connect(**postgres_connection_params)
        cursor = conn.cursor()
        
        cursor.execute("SELECT 1")
        result = cursor.fetchone()
        
        assert result[0] == 1
        
        cursor.close()
        conn.close()
    
    @pytest.mark.postgres
    def test_schema_exists(self, postgres_connection_params: dict):
        """Test that required schemas exist."""
        pytest.importorskip("psycopg2")
        
        import psycopg2
        
        conn = psycopg2.connect(**postgres_connection_params)
        cursor = conn.cursor()
        
        cursor.execute("""
            SELECT schema_name 
            FROM information_schema.schemata 
            WHERE schema_name IN ('bronze', 'silver', 'gold')
        """)
        
        schemas = [row[0] for row in cursor.fetchall()]
        
        assert 'silver' in schemas or len(schemas) >= 0  # May not exist in test env
        
        cursor.close()
        conn.close()
    
    @pytest.mark.postgres
    def test_insert_and_retrieve_user(self, postgres_connection_params: dict):
        """Test inserting and retrieving a user."""
        pytest.importorskip("psycopg2")
        
        from src.serving.postgres_loader import PostgresLoader
        from src.data_generator import AuctionDataGenerator
        
        loader = PostgresLoader(**postgres_connection_params)
        
        # Skip if database not properly initialized
        if not loader.health_check():
            pytest.skip("Database not available")
        
        # Generate test user
        generator = AuctionDataGenerator(seed=42)
        user = generator.generate_user()
        
        # This test would require the schema to be set up
        # In a real test environment, we'd have fixtures for this
        assert user is not None


class TestDataGeneratorIntegration:
    """Integration tests for data generation pipeline."""
    
    def test_generate_complete_dataset(self):
        """Test generating a complete dataset."""
        from src.data_generator import generate_sample_data
        
        users, items, bids, transactions = generate_sample_data(
            users=10,
            items=20,
            bids=100,
            seed=42,
        )
        
        assert len(users) == 10
        assert len(items) == 20
        assert len(bids) == 100
        
        # Verify referential integrity
        user_ids = {str(u.user_id) for u in users}
        item_ids = {str(i.item_id) for i in items}
        
        # All item sellers should be valid users
        for item in items:
            assert str(item.seller_id) in user_ids
        
        # All bids should reference valid items and users
        for bid in bids:
            assert str(bid.auction_id) in item_ids
            assert str(bid.bidder_id) in user_ids
    
    def test_data_serialization(self):
        """Test that all data types serialize correctly to JSON."""
        from src.data_generator import generate_sample_data
        
        users, items, bids, transactions = generate_sample_data(
            users=5,
            items=10,
            bids=20,
            seed=42,
        )
        
        # Test serialization
        for user in users:
            data = user.to_dict()
            json_str = json.dumps(data)
            assert json_str is not None
        
        for item in items:
            data = item.to_dict()
            json_str = json.dumps(data)
            assert json_str is not None
        
        for bid in bids:
            data = bid.to_dict()
            json_str = json.dumps(data)
            assert json_str is not None


class TestS3Integration:
    """Integration tests for S3 operations (with LocalStack)."""
    
    @pytest.mark.skipif(
        not os.getenv("AWS_ENDPOINT_URL"),
        reason="LocalStack not configured"
    )
    def test_bucket_operations(self):
        """Test S3 bucket operations with LocalStack."""
        from src.ingestion.s3_ingestion import S3Ingestion
        
        s3 = S3Ingestion(
            bucket_name="test-auction-bronze",
            endpoint_url=os.getenv("AWS_ENDPOINT_URL", "http://localhost:4566"),
        )
        
        # Create bucket
        result = s3.create_bucket()
        assert result is True
        
        # Check bucket exists
        assert s3.bucket_exists() is True
        
        # Upload test data
        test_data = {"test": "data", "timestamp": "2024-01-01T00:00:00"}
        result = s3.upload_data(test_data, "test/data.json")
        assert result is True
        
        # Read back data
        data = s3.read_json("test/data.json")
        assert data is not None
        assert data["test"] == "data"
        
        # List objects
        objects = s3.list_objects(prefix="test/")
        assert len(objects) > 0


class TestEndToEndPipeline:
    """End-to-end pipeline integration tests."""
    
    @pytest.mark.slow
    def test_data_generation_to_serialization(self):
        """Test full data generation and serialization flow."""
        from src.data_generator import AuctionDataGenerator
        from src.data_generator.schemas import BidEvent
        
        generator = AuctionDataGenerator(seed=42)
        
        # Generate entities
        users = generator.generate_users(5)
        items = generator.generate_items(10, ensure_active=5)
        
        # Generate bids and create events
        events = []
        for _ in range(20):
            bid = generator.generate_bid()
            event = BidEvent(payload=bid)
            events.append(event)
        
        # Verify events are serializable
        for event in events:
            data = event.to_dict()
            json_str = json.dumps(data)
            
            # Verify can deserialize
            parsed = json.loads(json_str)
            assert parsed["event_type"] == "bid_placed"
            assert "payload" in parsed
            assert "bid_id" in parsed["payload"]
