Auction-App-Data-Processing_AWS-Pipeline/requirements.txt
# =============================================================================
# Auction Data Pipeline - Python Dependencies
# =============================================================================
# Install with: pip install -r requirements.txt
# =============================================================================

# -----------------------------------------------------------------------------
# Core Data Processing
# -----------------------------------------------------------------------------
pyspark==3.5.0
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0

# -----------------------------------------------------------------------------
# Kafka Integration
# -----------------------------------------------------------------------------
kafka-python>=2.0.2
confluent-kafka>=2.3.0

# -----------------------------------------------------------------------------
# Database Connectivity
# -----------------------------------------------------------------------------
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
duckdb>=0.9.0

# -----------------------------------------------------------------------------
# AWS Integration
# -----------------------------------------------------------------------------
boto3>=1.34.0
botocore>=1.34.0

# -----------------------------------------------------------------------------
# Data Generation (Synthetic Data)
# -----------------------------------------------------------------------------
faker>=22.0.0
mimesis>=12.0.0

# -----------------------------------------------------------------------------
# Data Validation & Quality
# -----------------------------------------------------------------------------
pydantic>=2.5.0
great-expectations>=0.18.0

# -----------------------------------------------------------------------------
# Configuration & Environment
# -----------------------------------------------------------------------------
python-dotenv>=1.0.0
pyyaml>=6.0.0

# -----------------------------------------------------------------------------
# CLI & Utilities
# -----------------------------------------------------------------------------
click>=8.1.0
rich>=13.0.0
tqdm>=4.66.0

# -----------------------------------------------------------------------------
# Logging & Monitoring
# -----------------------------------------------------------------------------
structlog>=23.0.0

# -----------------------------------------------------------------------------
# HTTP & API
# -----------------------------------------------------------------------------
requests>=2.31.0
httpx>=0.26.0
